+++ Media Player SDK/WinForms/CSharp/Main Demo/Form1.cs
@@ -11,6 +11,7 @@ namespace Media_Player_Demo
     using System.Drawing.Imaging;
     using System.Globalization;
     using System.IO;
+    using System.Linq;
     using System.Threading.Tasks;
     using System.Windows.Forms;
     using VisioForge.Core;