+++ Video Edit SDK X (crossplatform)/Avalonia/Video Join Demo X AV/MainWindow.axaml.cs
@@ -221,86 +221,84 @@ namespace Video_Join_Demo_X_AV
             VideoEdit1.Debug_Telemetry = cbTelemetry.IsChecked == true;
 
             Log.Clear();
-
-            VideoEdit1.Output_Filename = edOutput.Text;
-
+            
             switch (cbOutputFormat.SelectedIndex)
             {
                 case 0:
                     {
-                        var mp4Output = new MP4Output();
+                        var mp4Output = new MP4Output(edOutput.Text);
                         VideoEdit1.Output_Format = mp4Output;
                         break;
                     }
                 case 1:
                     {
-                        var webmOutput = new WebMOutput();
+                        var webmOutput = new WebMOutput(edOutput.Text);
                         VideoEdit1.Output_Format = webmOutput;
                         break;
                     }
                 case 2:
                     {
-                        var aviOutput = new AVIOutput();
+                        var aviOutput = new AVIOutput(edOutput.Text);
                         VideoEdit1.Output_Format = aviOutput;
                         break;
                     }
                 case 3:
                     {
-                        var mkvOutput = new MKVOutput();
+                        var mkvOutput = new MKVOutput(edOutput.Text);
                         VideoEdit1.Output_Format = mkvOutput;
                         break;
                     }
                 case 4:
                     {
-                        var wmvOutput = new WMV1Output();
+                        var wmvOutput = new WMV1Output(edOutput.Text);
                         VideoEdit1.Output_Format = wmvOutput;
                         break;
                     }
                 case 5:
                     {
-                        var dvOutput = new DVOutput();
+                        var dvOutput = new DVOutput(edOutput.Text);
                         VideoEdit1.Output_Format = dvOutput;
                         break;
                     }
                 case 6:
                     {
-                        var acmOutput = new WAVOutput();
+                        var acmOutput = new WAVOutput(edOutput.Text);
                         VideoEdit1.Output_Format = acmOutput;
                         break;
                     }
                 case 7:
                     {
-                        var mp3Output = new MP3Output();
+                        var mp3Output = new MP3Output(edOutput.Text);
                         VideoEdit1.Output_Format = mp3Output;
                         break;
                     }
                 case 8:
                     {
-                        var m4aOutput = new M4AOutput();
+                        var m4aOutput = new M4AOutput(edOutput.Text);
                         VideoEdit1.Output_Format = m4aOutput;
                         break;
                     }
                 case 9:
                     {
-                        var wmaOutput = new WMA1Output();
+                        var wmaOutput = new WMA1Output(edOutput.Text);
                         VideoEdit1.Output_Format = wmaOutput;
                         break;
                     }
                 case 10:
                     {
-                        var oggVorbisOutput = new OGGVorbisOutput();
+                        var oggVorbisOutput = new OGGVorbisOutput(edOutput.Text);
                         VideoEdit1.Output_Format = oggVorbisOutput;
                         break;
                     }
                 case 11:
                     {
-                        var flacOutput = new FLACOutput();
+                        var flacOutput = new FLACOutput(edOutput.Text);
                         VideoEdit1.Output_Format = flacOutput;
                         break;
                     }
                 case 12:
                     {
-                        var speexOutput = new SpeexOutput();
+                        var speexOutput = new SpeexOutput(edOutput.Text);
                         VideoEdit1.Output_Format = speexOutput;
                         break;
                     }