+++ Video Edit SDK/WinForms/CSharp/Main Demo/Form1.cs
@@ -18,6 +18,7 @@ namespace VideoEdit_CS_Demo
     using System.Drawing;
     using System.Globalization;
     using System.IO;
+    using System.Linq;
     using System.Threading.Tasks;
     using System.Windows.Forms;
     using VisioForge.Core.Helpers;